+++ CHANGELOG.md
@@ -16,6 +16,24 @@ Types of changes:
 
 ## [Unreleased]
 
+## [0.4.0] - 2022-01-25
+
+### Changed
+- Better Kubernetes support
+- Uses new version of ecHome Python SDK
+
+## [0.3.0] - 2021-11-12
+
+### Added
+- Support for new endpoints with updated python-sdk library
+- Some commands now have an extra flag for producing wider table output with more information
+
+### Changed
+- Help/Usage output is more consolidated and easier to read
+- Consolidated commands to match the endpoints
+- The format argument (--format, -f) has been changed to --output, -o. The options have stayed the same
+
+
 ## [0.2.0] - 2021-04-03
 
 ### Fixed 
@@ -35,6 +53,9 @@ Types of changes:
 ### Added
 - Initial import of ecHome CLI
 
-[unreleased]: https://github.com/mgtrrz/echome-cli/compare/0.1.1...HEAD
+[unreleased]: https://github.com/mgtrrz/echome-cli/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/mgtrrz/echome-cli/compare/0.3.0...0.4.0
+[0.3.0]: https://github.com/mgtrrz/echome-cli/compare/0.2.0...0.3.0
+[0.2.0]: https://github.com/mgtrrz/echome-cli/compare/0.1.0...0.2.0
 [0.1.1]: https://github.com/mgtrrz/echome-cli/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/mgtrrz/echome-cli/releases/tag/0.1.0